#1 Restructure to standard yakworks projects structure
