+++ build.gradle
@@ -1,70 +1,65 @@
 buildscript {
-    repositories {
-        mavenLocal()
-        maven { url "https://repo.grails.org/grails/core" }
-    }
-    dependencies {
-        classpath "org.grails:grails-gradle-plugin:$grailsVersion"
-    }
+	repositories {
+		mavenLocal()
+		maven { url "https://repo.grails.org/grails/core" }
+	}
+	dependencies {
+		classpath "org.grails:grails-gradle-plugin:$grailsVersion"
+		classpath "org.grails:grails-docs:$grailsVersion"
+		classpath 'com.jfrog.bintray.gradle:gradle-bintray-plugin:1.7.3'
+		classpath 'io.github.groovylang.groovydoc:groovydoc-gradle-plugin:1.0.1'
+		classpath "com.bertramlabs.plugins:asset-pipeline-gradle:2.8.2"
+	}
 }
 
-version "0.4"
-group "org.grails.plugins"
-
-
-apply plugin:"idea"
-apply plugin:"org.grails.grails-plugin"
-apply plugin:"org.grails.grails-plugin-publish"
-apply plugin:"org.grails.grails-gsp"
+apply plugin: 'idea'
+apply plugin:'groovy'
 
 repositories {
-    mavenLocal()
-    maven { url "https://repo.grails.org/grails/core" }
+	mavenLocal()
+	maven { url "https://repo.grails.org/grails/core" }
 }
 
-dependencies {
-    compile "org.springframework.boot:spring-boot-starter-logging"
-    compile "org.springframework.boot:spring-boot-autoconfigure"
-    compile "org.grails:grails-core"
-    compile "org.springframework.boot:spring-boot-starter-actuator"
-    provided "org.springframework.boot:spring-boot-starter-tomcat"
-    compile "org.grails:grails-dependencies"
-    compile "org.grails:grails-web-boot"
-    compile "org.grails.plugins:cache"
-    compile "org.grails.plugins:scaffolding"
-    console "org.grails:grails-console"
-    profile "org.grails.profiles:web-plugin"
-    provided "org.grails:grails-plugin-services"
-    provided "org.grails:grails-plugin-domain-class"
+//Root project and view-tools plugin uses same version.
+version project.projectVersion
 
-    compile 'org.grails.plugins:quartz:2.0.12'
-
-    testCompile "org.grails:grails-plugin-testing"
-
+project(':quartz-config-scheduler') {
+	version = project.projectVersion
 }
 
-bootRun {
-    jvmArgs('-Dspring.output.ansi.enabled=always')
-    addResources = true
+ext {
+	commonBuild = "https://raw.githubusercontent.com/yakworks/grails-common-build/master"
+	grailsApps = []
+	grailsPlugins = ['quartz-config-scheduler']
 }
-// enable if you wish to package this plugin as a standalone application
-bootRepackage.enabled = false
 
 
-grailsPublish {
-    user = System.getenv("BINTRAY_USER") ?: project.bintrayUser
-    key = System.getenv("BINTRAY_KEY") ?: project.bintrayKey
+subprojects {
 
-    desc = 'Allow creating quartz job from configuration'
-    developers = [snimavat: 'Sudhir Nimavat']
-    githubSlug = '9ci/quartz-config-schedular'
-    issueTrackerUrl = 'https://github.com/9ci/quartz-config-scheduler/issues'
-    vcsUrl = 'https://github.com/9ci/quartz-config-schedular'
-    license {
-        name = 'Apache-2.0'
-    }
-    title = 'Quartz Config Scheduler Plugin'
-    //userOrg = '9ci' //the Org to publish this under in
-    repo = 'grails-plugins' //the repository you setup in bintray
-    userOrg = '9ci'
+	boolean isGrailsApp = grailsApps.contains(project.name)
+	boolean isGrailsPlugin = grailsPlugins.contains(project.name)
+	boolean isGrailsProject = isGrailsApp || isGrailsPlugin
+	boolean isBuildSnapshot = version.toString().endsWith("-SNAPSHOT")
+
+	println project.name + ":" + version.toString()
+
+	ext {
+		userOrg = "grails"
+		set "tomcat.version", "8.5.5"
+		isSnapshot = isBuildSnapshot
+	}
+
+	apply from:"$commonBuild/common.gradle"
+
+	if (isGrailsPlugin) {
+		apply from: "${commonBuild}/common-plugin.gradle"
+		apply from: "${commonBuild}/common-publishing.gradle"
+		group "org.grails.plugins"
+	}
+
+	if (isGrailsApp) {
+		apply from: "${commonBuild}/common-project.gradle"
+		group "org.grails"
+	}
+
 }
