+++ gradle.properties
@@ -1,3 +1,10 @@
 grailsVersion=3.2.11
 gormVersion=6.0.12.RELEASE
 gradleWrapperVersion=3.4.1
+projectVersion=0.5-SNAPSHOT
+
+title=Quartz Config Scheduler
+projectDesc=Allow creating quartz job from configuration
+projectUrl=https://github.com/yakworks/quartz-config-scheduler
+githubSlug=yakworks/quartz-config-scheduler
+
