--- grails-app/controllers/grails/plugin/quartzconfigscheduler/UrlMappings.groovy
@@ -1,16 +1,0 @@
-package grails.plugin.quartzconfigscheduler
-
-class UrlMappings {
-
-    static mappings = {
-        "/$controller/$action?/$id?(.$format)?"{
-            constraints {
-                // apply constraints here
-            }
-        }
-
-        "/"(view:"/index")
-        "500"(view:'/error')
-        "404"(view:'/notFound')
-    }
-}
