--- grails-app/init/grails/plugin/quartzconfigscheduler/BootStrap.groovy
@@ -1,9 +1,0 @@
-package grails.plugin.quartzconfigscheduler
-
-class BootStrap {
-
-    def init = { servletContext ->
-    }
-    def destroy = {
-    }
-}
