+++ plugin/build.gradle
@@ -1,0 +1,70 @@
+buildscript {
+    repositories {
+        mavenLocal()
+        maven { url "https://repo.grails.org/grails/core" }
+    }
+    dependencies {
+        classpath "org.grails:grails-gradle-plugin:$grailsVersion"
+    }
+}
+
+version "0.4"
+group "org.grails.plugins"
+
+
+apply plugin:"idea"
+apply plugin:"org.grails.grails-plugin"
+apply plugin:"org.grails.grails-plugin-publish"
+apply plugin:"org.grails.grails-gsp"
+
+repositories {
+    mavenLocal()
+    maven { url "https://repo.grails.org/grails/core" }
+}
+
+dependencies {
+    compile "org.springframework.boot:spring-boot-starter-logging"
+    compile "org.springframework.boot:spring-boot-autoconfigure"
+    compile "org.grails:grails-core"
+    compile "org.springframework.boot:spring-boot-starter-actuator"
+    provided "org.springframework.boot:spring-boot-starter-tomcat"
+    compile "org.grails:grails-dependencies"
+    compile "org.grails:grails-web-boot"
+    compile "org.grails.plugins:cache"
+    compile "org.grails.plugins:scaffolding"
+    console "org.grails:grails-console"
+    profile "org.grails.profiles:web-plugin"
+    provided "org.grails:grails-plugin-services"
+    provided "org.grails:grails-plugin-domain-class"
+
+    compile 'org.grails.plugins:quartz:2.0.12'
+
+    testCompile "org.grails:grails-plugin-testing"
+
+}
+
+bootRun {
+    jvmArgs('-Dspring.output.ansi.enabled=always')
+    addResources = true
+}
+// enable if you wish to package this plugin as a standalone application
+bootRepackage.enabled = false
+
+
+grailsPublish {
+    user = System.getenv("BINTRAY_USER") ?: project.bintrayUser
+    key = System.getenv("BINTRAY_KEY") ?: project.bintrayKey
+
+    desc = 'Allow creating quartz job from configuration'
+    developers = [snimavat: 'Sudhir Nimavat']
+    githubSlug = '9ci/quartz-config-schedular'
+    issueTrackerUrl = 'https://github.com/9ci/quartz-config-scheduler/issues'
+    vcsUrl = 'https://github.com/9ci/quartz-config-schedular'
+    license {
+        name = 'Apache-2.0'
+    }
+    title = 'Quartz Config Scheduler Plugin'
+    //userOrg = '9ci' //the Org to publish this under in
+    repo = 'grails-plugins' //the repository you setup in bintray
+    userOrg = '9ci'
+}
