+++ plugin/grails-app/init/grails/plugin/quartzconfigscheduler/Application.groovy
