+++ plugin/src/main/groovy/grails/plugin/quartzconfigscheduler/QuartzConfigSchedulerGrailsPlugin.groovy
@@ -1,0 +1,9 @@
+package grails.plugin.quartzconfigscheduler
+
+import grails.plugins.*
+
+class QuartzConfigSchedulerGrailsPlugin extends Plugin {
+
+    def profiles = ['web']
+
+}
