+++ plugin/src/main/resources/META-INF/spring.factories
